+++ src/common/utils.ts
@@ -11,51 +11,34 @@ export function debounce<T extends (...args: any[]) => void>(fn: T, delayMs = 25
 }
 
 /**
- * Fires only when a real text editor becomes active, excludes Output, Debug
+ * Fires only when a main-area text editor becomes active, excludes Output, Debug
  * Console, and other non-document views.
  */
-export function onSafeActiveTextEditorChanged(
+export function onMainAreaActiveTextEditorChanged(
   callback: (editor: vscode.TextEditor | undefined) => void,
 ): vscode.Disposable {
-  return vscode.window.tabGroups.onDidChangeTabGroups(({ changed }) => {
-    if (changed.length === 1 && changed[0].isActive) {
-      callback(safeActiveTextEditor());
+  return vscode.window.onDidChangeActiveTextEditor((editor) => {
+    if (editor?.viewColumn) {
+      callback(editor);
     }
   });
 }
 
 /**
- * Returns the active text editor only if it's a real file editor, excludes
- * Output, Debug Console, and other non-document views.
- */
-export function safeActiveTextEditor(): vscode.TextEditor | undefined {
-  const uri = safeActiveTextEditorUri();
-  return (
-    uri &&
-    vscode.window.visibleTextEditors.find(
-      (editor) => editor.document.uri.toString() === uri.toString(),
-    )
-  );
-}
-
-/**
- * Returns the URI of the active text editor only if it's a real file editor,
+ * Returns the URI of the active text editor only if it's a main-area file editor,
  * excludes Output, Debug Console, and views.
  */
-export function safeActiveTextEditorUri(): vscode.Uri | undefined {
+export function mainAreaActiveTextEditorUri(): vscode.Uri | undefined {
+  let uri: vscode.Uri | undefined;
+
   const tab = vscode.window.tabGroups.activeTabGroup.activeTab;
-
-  if (
-    tab &&
-    (tab.input instanceof vscode.TabInputText || tab.input instanceof vscode.TabInputTextDiff)
-  ) {
-    const uri =
-      tab.input instanceof vscode.TabInputText
-        ? tab.input.uri
-        : (tab.input.modified ?? tab.input.original);
-
-    return uri;
+  if (tab) {
+    if (tab.input instanceof vscode.TabInputText) {
+      uri = tab.input.uri;
+    } else if (tab.input instanceof vscode.TabInputTextDiff) {
+      uri = tab.input.modified ?? tab.input.original;
+    }
   }
 
-  return undefined;
+  return uri;
 }